Changed client attribute code sample to client labels.
