+++ clients/client-labels/src/main/java/ProgrammaticAttributesExample.java
@@ -1,28 +1,12 @@
-/*
- * Copyright (c) 2008-2018, Hazelcast, Inc. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
 import com.hazelcast.client.HazelcastClient;
+import com.hazelcast.client.config.ClientConfig;
 import com.hazelcast.config.Config;
 import com.hazelcast.config.JoinConfig;
 import com.hazelcast.core.ClientListener;
 import com.hazelcast.core.Hazelcast;
 import com.hazelcast.core.HazelcastInstance;
 
-public class DeclarativeAttributesExample {
-
+public class ProgrammaticAttributesExample {
     public static void main(String[] args) {
         Config config = new Config();
         JoinConfig join = config.getNetworkConfig().getJoin();
@@ -33,7 +17,7 @@ public class DeclarativeAttributesExample {
             @Override
             public void clientConnected(com.hazelcast.core.Client client) {
                 System.out.println("Client : " + client.getName() + " is connected to member: " + instance.getName());
-                System.out.println("Client : " + client.getName() + " is connected with attributes " + client.getLabels());
+                System.out.println("Client : " + client.getName() + " is connected with labels " + client.getLabels());
 
             }
 
@@ -43,7 +27,12 @@ public class DeclarativeAttributesExample {
             }
         });
 
-        //client opened with xml config. See ./resources/hazelcast-client.xml
-        HazelcastClient.newHazelcastClient();
+        ClientConfig clientConfig = new ClientConfig();
+        //setting instance name and attributes
+        clientConfig.setInstanceName("SampleClientName");
+        clientConfig.addLabel("role admin");
+        clientConfig.addLabel("region foo");
+
+        HazelcastClient.newHazelcastClient(clientConfig);
     }
 }
